refactor(kubernetes): move crypto policy helpers
